Backend_DRF/requirements.txt
Django==3.2.14
django-filter==21.1
djangorestframework==3.13.1
PyMySQL==1.0.2
django-cors-headers==3.11.0
cryptography==36.0.2